Description
STEP 1 Empathize, Research: The electricity tariff is the amount charged for the supply of electricity usage. The charges of electricity cost per unit in Mumbai vary for different customers depending on the season or time of the day when the energy is used. In simple words, Electricity tariff is a method through which the customer is charged a particular amount for energy power consumption. This component of the electricity bill (tariff) includes the cost of supplying and producing the electricity at a reasonable price.

The types of tariffs that the customer will eventually end up paying are highly dependent on electricity consumption, as consumers’ requirements can vary. It usually starts with LT (low tension) or HT (high tension). LT is generally applied for more residential and commercial areas where energy usage is comparatively low. However, HT codes are used for larger complexes and industries. Bigger consumers pay higher power tariffs as their energy needs are much higher than those of other consumers, including domestic and commercial consumers.

The various factors on which the electricity tariffs are dependent upon:

The type of the load
The power of the load
The time when the maximum load is required
The total amount of used electric energy
https://www.adanielectricity.com/tariff

 

STEP 2 DEFINE: State user needs : Previously the users felt difficult to calculate the bill of industries but now it can be calculated easily with the help of programs.


 

 

STEP 3 Ideate : The main purpose of this c program was to find the electricity bill consumed by industries. So we added the input to count number of machines, fans, lights and the usage per hour.  This program also allows user to fix the electricity rate for one unit. It also shows electricity consumption of individual devices with the help of which the user can identify which device is being consumed more. At last it shows total bill.

 

Algorithm: Step 1: Start

 

Step 2: Input the rate of electricity as 5.00 per unit.

 

Step 3: Input the no of hours.

 

Step 4: Input the no of machines, fans, lights etc.

 

Step 5: Input the power rating of each equipment.

 

Step 6: units consumed = (power consumed*hours)/1000

 

Step 7:  Bill= units consumed*rate

 

Step 8: Total bill = Bill of machines +Bill of fans +Bill of lights.

 

Step 9: End

 

 

STEP 4 Prototype:  In this program we calculate unit consumed= (power consumed*hours)/1000. Then we calculate bill by multiplying it with rate.

 

 

STEP 5 TESTING: We made the program to calculate the bill for industries. We faced some difficulties which we solved with the help of google.